Implement new_from_where() to avoid double selects
in some places in the code, we follow the pattern of
select all the ids that match a query, and then making a second query
on that list of IDs. This is wasteful when we can just ask for the data up
front.

This commit implements a new_from_where() method on Bugzilla::Object
in terms of the existing _do_list_select().