Rationalize the match() code in the tracking flags extension
Currently the match code is very confusing. It was copied from Bugzilla::Bug,
and implements by-name and by-object criteria for the match() call.

I initially noticed this because at some point we send a Bugzilla::Component
object as if it was a string name value to mysql. This seems to be harmless
but is clearly an error.

To rationalize this code, I have removed the by-name magic, and instead
accept either ids or objects.