Bug 1352264 - Preload all templates
This commit adds a new Template::Provider subclass that stores all templates in
memory.
This cuts down on calls to stat() as well calls relating to managing path names
(File::Spec->splitpath and so on).

It also uses less memory overall, as the memory required to store all templates
is shared between the httpd procceses.

The traditional provider is used when not running under mod_perl.