use state var for serialization keys
As far as I understand it, serialization keys exist to track what fields are
"native" to an object and which ones need to be ignored / not cached.

I see no reason for this to be per-request, so it can be per-process,
and a state variable is a good way of doing this.